rename `app` config to `internal` config
