+++ lib/configs/internal.js
