+++ lib/index.js
@@ -15,7 +15,7 @@ module.exports = {
     'unescaped-html-literal': require('./rules/unescaped-html-literal')
   },
   configs: {
-    app: require('./configs/app'),
+    internal: require('./configs/internal'),
     browser: require('./configs/browser'),
     recommended: require('./configs/recommended'),
     typescript: require('./configs/typescript')
